requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.8.0
statsmodels>=0.13.0
alpha_vantage>=2.3.1
python-dotenv>=0.19.0
matplotlib>=3.4.3
seaborn>=0.11.2
ta>=0.7.0  # Technical analysis library
jupyter>=1.0.0
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
flask>=2.0.1
flask-cors>=3.0.10
plotly>=5.3.1
dash>=2.0.0
dash-bootstrap-components>=1.0.0
gunicorn>=20.1.0 